got the majority of namespace-related things working.
managed to find a way to transfer control to the normal codegen code after the first non-namespace expression in the flattened dot-operator tree. Still need to implement that for functions. Problem is, right now this entire thing only works when it is not the "initVal" of an inferred variable, as that requires getting the type of the expression chain, which is non-trivial. Right now, I handled it by simply screwing side-effects, and calling codegen multiple times. This will be disastrous if there are any function calls in the expression chain, which will then be codegened multiple times. finding a fix.