fortunately it works if we use the empty-module thing.
actually, i'm not sure if we're facing the Types-across-module bug, but just to be sure i added an external map that tracks types that are created, separately from the Module -- presumably the module adds a named type to its list if it encounters a Value with that type? idk. still works though


Former-commit-id: 53f4277577cd4e5c81bfda11736c70a98de7f63b