globals work now, but namespacing is a little iffy. For something like this: ``` namespace foo { 	namespace bar 	{ 	}
	fn somefn() { }
}
```

You can call `somefn` by using `foo.bar.somefn()`, as well as `foo.somefn()`. This is probably due to how we traverse upwards when looking for the definition of something; we need a way to disable that when we're using a dot-op, because that implicit upward searching should be disabled since there's an explicit scope provided. If anything, only the very first member -- `foo` in this case -- should be searched upwards, since we could be in an inner namespace. Anything after that should be a one-level search.