fix #47
limit codegen-caching to Defns; those still need to be cached
because of the way the rest of the compiler is written (ie. you
should be able to unconditionally call `->codegen()` on your
target [eg. a type, a vardefn, or a funcdefn] without doing any
checks).

for the rest of the statements, caching shouldn't be done, since
that was causing the defer problem (the deferred block/statement
was already codegened and cached, even though it needded to be
emitted more than once)